+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added a handler (`/layers/workflowId/{provider}/{layer}`) to register a layer by id and return the resulting workflow id.
+- Added a handler (`/layers/{provider}/{layer}/workflowId`) to register a layer by id and return the resulting workflow id.
 
   - https://github.com/geo-engine/geoengine/pull/692